Add Uncertainty Checking, Foundations of ResetButton, State comparing
Theres now a check for when the player hovers over a tile, if its covered then the Game State will switch to "Uncertain". This is a component of ResetButton, which now has its textures added into the game, although the class itself does not exist yet. Some enchancements have been made to TileState and GameState, including dedicated comparing functions