Add Configuration Support/Strip down ConfigMenu
Add a configuration mode that allows the current game mode/game information to be stored in a file that can be read later. This also adds difficulties [As integers] into the game. Also reduce ConfigMenu again to a simple blank pane, as the original idea of the settings menu was not working out.